CRUD API Posts

+++ app/Http/Controllers/API/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\API;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Http\JsonResponse;
 
 use App\Models\Category;
 use App\Http\Requests\Api\Category\{
@@ -12,28 +13,32 @@ use App\Http\Requests\Api\Category\{
 
 class CategoryController extends Controller
 {
-    public function index()
+    public function index(): JsonResponse
     {
         return response()->json(Category::paginate(10));
     }
 
-    public function store(StoreRequest $request)
+    public function store(StoreRequest $request): JsonResponse
     {
         return response()->json(Category::create($request->validated()));
     }
 
-    public function show(Category $category)
+    public function show(Category $category): JsonResponse
     {
         return response()->json($category);
     }
 
-    public function update(UpdateRequest $request, Category $category)
+    public function update(UpdateRequest $request, Category $category): JsonResponse
     {
         return response()->json($category->update($request->validated()));
     }
 
-    public function destroy(Category $category)
+    public function destroy(Category $category): JsonResponse
     {
+        if ($category->posts()->count() > 0) {
+            return response()->json('A category can\'t be deleted because it has some posts');
+        }
+
         $category->delete();
         return response()->json('Category deleted successfully');
     }
